refactor!: remove v1 vault implementation
